+++ packages/media/media-card/src/root/__tests__/unit/cardLoader.spec.tsx
@@ -65,8 +65,13 @@ describe('Async Card Loader', () => {
   });
 
   describe('When the async import returns with success', () => {
+    let MediaPickerAnalyticsErrorBoundary: React.ReactComponentElement<any>;
     beforeEach(() => {
       jest.unmock('../../card');
+      jest.unmock('../../media-card-analytics-error-boundary');
+      MediaPickerAnalyticsErrorBoundary = jest.requireActual(
+        '../../media-card-analytics-error-boundary',
+      ).default;
     });
 
     it('should render Card component', async () => {
@@ -78,6 +83,35 @@ describe('Async Card Loader', () => {
       await nextTick();
 
       expect(wrapper.state().Card).not.toBeUndefined();
+    });
+
+    it('should render Error boundary component', async () => {
+      const wrapper = await mount<
+        CardWithMediaClientConfigProps,
+        AsyncCardState
+      >(<CardLoader {...props} />);
+      await nextTick();
+      expect(wrapper.find(MediaPickerAnalyticsErrorBoundary)).toBeDefined();
+    });
+  });
+
+  describe('When the async import for Error Boundary returns with error', () => {
+    beforeEach(() => {
+      jest.unmock('../../card');
+      jest.mock('../../media-card-analytics-error-boundary', () => {
+        throw new Error('Forcing error boundary async import error');
+      });
+    });
+
+    it('should render CardLoading component', async () => {
+      const wrapper = await mount<
+        CardWithMediaClientConfigProps,
+        AsyncCardState
+      >(<CardLoader {...props} />);
+
+      await nextTick();
+
+      expect(wrapper.find(CardLoading)).toHaveLength(1);
     });
   });
 });