+++ packages/media/media-card/src/root/card/cardLoader.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { WithContextOrMediaClientConfigProps } from '@atlaskit/media-client';
 import { CardProps, CardLoading } from '../..';
+import { MediaCardAnalyticsErrorBoundaryProps } from '../media-card-analytics-error-boundary';
 
 export type CardWithMediaClientConfigProps = WithContextOrMediaClientConfigProps<
   CardProps
@@ -9,8 +10,13 @@ type CardWithMediaClientConfigComponent = React.ComponentType<
   CardWithMediaClientConfigProps
 >;
 
+type MediaCardErrorBoundaryComponent = React.ComponentType<
+  MediaCardAnalyticsErrorBoundaryProps
+>;
+
 export interface AsyncCardState {
   Card?: CardWithMediaClientConfigComponent;
+  MediaCardErrorBoundary?: MediaCardErrorBoundaryComponent;
 }
 
 export default class CardLoader extends React.PureComponent<
@@ -19,21 +25,34 @@ export default class CardLoader extends React.PureComponent<
 > {
   static displayName = 'AsyncCard';
   static Card?: CardWithMediaClientConfigComponent;
+  static MediaCardErrorBoundary?: MediaCardErrorBoundaryComponent;
 
   state: AsyncCardState = {
     Card: CardLoader.Card,
+    MediaCardErrorBoundary: CardLoader.MediaCardErrorBoundary,
   };
 
   async componentDidMount() {
     if (!this.state.Card) {
       try {
-        const [mediaClient, cardModule] = await Promise.all([
+        const [
+          mediaClient,
+          cardModule,
+          mediaCardErrorBoundaryModule,
+        ] = await Promise.all([
           import(/* webpackChunkName:"@atlaskit-media-client" */ '@atlaskit/media-client'),
           import(/* webpackChunkName:"@atlaskit-internal_Card" */ './index'),
+          import(/* webpackChunkName:"@atlaskit-internal_MediaCardErrorBoundary" */ '../media-card-analytics-error-boundary'),
         ]);
 
         CardLoader.Card = mediaClient.withMediaClient(cardModule.Card);
-        this.setState({ Card: CardLoader.Card });
+        CardLoader.MediaCardErrorBoundary =
+          mediaCardErrorBoundaryModule.default;
+
+        this.setState({
+          Card: CardLoader.Card,
+          MediaCardErrorBoundary: CardLoader.MediaCardErrorBoundary,
+        });
       } catch (error) {
         // TODO [MS-2278]: Add operational error to catch async import error
       }
@@ -42,11 +61,16 @@ export default class CardLoader extends React.PureComponent<
 
   render() {
     const { dimensions } = this.props;
+    const { Card, MediaCardErrorBoundary } = this.state;
 
-    if (!this.state.Card) {
+    if (!Card || !MediaCardErrorBoundary) {
       return <CardLoading dimensions={dimensions} />;
     }
 
-    return <this.state.Card {...this.props} />;
+    return (
+      <MediaCardErrorBoundary>
+        <Card {...this.props} />
+      </MediaCardErrorBoundary>
+    );
   }
 }