+++ packages/media/media-card/src/root/media-card-analytics-error-boundary.tsx
@@ -2,16 +2,16 @@ import React from 'react';
 
 import AnalyticsErrorBoundary from '@atlaskit/analytics-next/AnalyticsErrorBoundary';
 
-export type MediaPickerAnalyticsErrorBoundaryProps = {
+export type MediaCardAnalyticsErrorBoundaryProps = {
   data?: { [k: string]: any };
 };
 
 export const ANALYTICS_MEDIA_CHANNEL = 'media';
 
-export default class MediaPickerAnalyticsErrorBoundary extends React.Component<
-  MediaPickerAnalyticsErrorBoundaryProps
+export default class MediaCardAnalyticsErrorBoundary extends React.Component<
+  MediaCardAnalyticsErrorBoundaryProps
 > {
-  static displayName = 'MediaPickerAnalyticsErrorBoundary';
+  static displayName = 'MediaCardAnalyticsErrorBoundary';
 
   render() {
     const { data = {}, children } = this.props;