+++ packages/media/media-card/src/utils/cardActions/cardActionsDropdownMenu.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Component } from 'react';
+import { Component, HTMLAttributes } from 'react';
 
 import MoreIcon from '@atlaskit/icon/glyph/more';
 import DropdownMenu, {
@@ -9,12 +9,58 @@ import DropdownMenu, {
 
 import { CardAction } from '../../actions';
 import { CardActionButton } from './styled';
+import {
+  withAnalyticsEvents,
+  WithAnalyticsEventsProps,
+} from '@atlaskit/analytics-next';
+import { createAndFireMediaEvent } from '../analytics';
 
 export type CardActionsDropdownMenuProps = {
   readonly actions: CardAction[];
 
   readonly triggerColor?: string;
   readonly onOpenChange?: (attrs: { isOpen: boolean }) => void;
+};
+
+type CardActionButtonProps = HTMLAttributes<HTMLDivElement> &
+  WithAnalyticsEventsProps;
+const CardActionButtonWithProps = (props: CardActionButtonProps) => (
+  <CardActionButton {...props} />
+);
+
+const CardActionButtonWithAnalytics = withAnalyticsEvents({
+  onClick: createAndFireMediaEvent({
+    eventType: 'ui',
+    action: 'clicked',
+    actionSubject: 'button',
+    actionSubjectId: 'mediaCardDropDownMenu',
+  }),
+})(CardActionButtonWithProps);
+
+type DropdownItemProps = any & WithAnalyticsEventsProps; // Trick applied due to the lack of props type of DropdownItem
+const DropdownItemWithProps = (props: DropdownItemProps) => (
+  <DropdownItem {...props} />
+);
+
+const createDropdownItemWithAnalytics = (action: CardAction, index: number) => {
+  const { label, handler } = action;
+  const DropdownItemWithAnalytics = withAnalyticsEvents({
+    onClick: createAndFireMediaEvent({
+      eventType: 'ui',
+      action: 'clicked',
+      actionSubject: 'button',
+      actionSubjectId: 'mediaCardDropDownMenuItem',
+      attributes: {
+        label,
+      },
+    }),
+  })(DropdownItemWithProps);
+
+  return (
+    <DropdownItemWithAnalytics key={index} onClick={handler}>
+      {label}
+    </DropdownItemWithAnalytics>
+  );
 };
 
 export class CardActionsDropdownMenu extends Component<
@@ -28,17 +74,13 @@ export class CardActionsDropdownMenu extends Component<
         <DropdownMenu
           onOpenChange={onOpenChange}
           trigger={
-            <CardActionButton style={{ color: triggerColor }}>
+            <CardActionButtonWithAnalytics style={{ color: triggerColor }}>
               <MoreIcon label="more" />
-            </CardActionButton>
+            </CardActionButtonWithAnalytics>
           }
         >
           <DropdownItemGroup>
-            {actions.map(({ label, handler }, index) => (
-              <DropdownItem key={index} onClick={handler}>
-                {label}
-              </DropdownItem>
-            ))}
+            {actions.map(createDropdownItemWithAnalytics)}
           </DropdownItemGroup>
         </DropdownMenu>
       );