+++ packages/media/media-viewer/src/components/media-viewer-loader.tsx
@@ -53,9 +53,12 @@ export default class AsyncMediaViewer extends React.PureComponent<
           mediaViewerModule.MediaViewer,
         );
         AsyncMediaViewer.MediaViewer = MediaViewerWithClient;
+        AsyncMediaViewer.MediaViewerErrorBoundary =
+          mediaViewerErrorBoundaryModule.default;
+
         this.setState({
           MediaViewer: MediaViewerWithClient,
-          MediaViewerErrorBoundary: mediaViewerErrorBoundaryModule.default,
+          MediaViewerErrorBoundary: AsyncMediaViewer.MediaViewerErrorBoundary,
         });
       } catch (error) {
         // TODO [MS-2277]: Add operational error to catch async import error
@@ -64,16 +67,17 @@ export default class AsyncMediaViewer extends React.PureComponent<
   }
 
   render() {
-    if (!this.state.MediaViewer || !this.state.MediaViewerErrorBoundary) {
+    const { MediaViewer, MediaViewerErrorBoundary } = this.state;
+    if (!MediaViewer || !MediaViewerErrorBoundary) {
       return (
         <ModalSpinner blankedColor={colors.DN30} invertSpinnerColor={true} />
       );
     }
 
     return (
-      <this.state.MediaViewerErrorBoundary>
-        <this.state.MediaViewer {...this.props} />
-      </this.state.MediaViewerErrorBoundary>
+      <MediaViewerErrorBoundary>
+        <MediaViewer {...this.props} />
+      </MediaViewerErrorBoundary>
     );
   }
 }