+++ packages/media/smart-card/src/state/actions/index.ts
@@ -13,6 +13,8 @@ import {
   ACTION_RESOLVING,
   ACTION_RESOLVED,
   ACTION_ERROR,
+  ERROR_MESSAGE_OAUTH,
+  ERROR_MESSAGE_FATAL,
 } from './constants';
 import { CardAppearance } from '../../view/Card';
 import {
@@ -30,6 +32,7 @@ import {
   screenAuthPopupEvent,
 } from '../../utils/analytics';
 import { useSmartLinkContext } from '../context';
+import { JsonLd } from '../../client/types';
 
 export function useSmartCardActions(
   url: string,
@@ -77,65 +80,91 @@ export function useSmartCardActions(
     // has come back asynchronously, dispatch the resolved action for the card.
     if (isReloading || hasExpired || !hasData) {
       try {
-        const response = await connections.client.fetchData(resourceUrl);
-        const nextDefinitionId = getDefinitionId(response);
-        const nextStatus = getStatus(response);
-
-        // Dispatch analytics.
-        if (nextStatus === 'resolved') {
-          dispatchAnalytics(resolvedEvent(nextDefinitionId));
-        } else {
-          dispatchAnalytics(unresolvedEvent(nextStatus, nextDefinitionId));
-          // If we require authorization & do not have an authFlow available,
-          // throw an error and render as a normal blue link.
-          if (nextStatus === 'unauthorized' && config.authFlow !== 'oauth2') {
-            throw Error('Provider.authFlow is not set to OAuth2.');
-          }
-        }
-
-        dispatch(cardAction(ACTION_RESOLVED, { url: resourceUrl }, response));
+        /**
+         * There was a bizarre error in Android where the Webview would crash when
+         * we were using the syntax:
+         *    const response = await connections.client.fetchData(resourceUrl)
+         * Please DO NOT change this from a promise
+         * https://product-fabric.atlassian.net/browse/FM-2240
+         */
+        return connections.client.fetchData(resourceUrl).then(
+          response => {
+            handleResolvedLinkResponse(resourceUrl, response);
+          },
+          error => {
+            handleResolvedLinkError(resourceUrl, error);
+          },
+        );
       } catch (error) {
-        // Handle FatalErrors (completely failed to resolve data).
-        if (error.kind === 'fatal') {
-          // If there's no previous data in the store for this URL, then bail
-          // out and let the editor handle fallbacks (returns to a blue link).
-          if (!details || status !== 'resolved') {
-            throw error;
-          }
-
-          // If we already have resolved data for this URL in the store, then
-          // simply fallback to the previous data.
-          if (status === 'resolved') {
-            dispatch(
-              cardAction(ACTION_RESOLVED, { url: resourceUrl }, details),
-            );
-          }
-          // Handle AuthErrors (user did not have access to resource).
-        } else if (error.kind === 'auth') {
-          dispatch(
-            cardAction(
-              ACTION_RESOLVED,
-              { url: resourceUrl },
-              {
-                meta: {
-                  visibility: 'restricted',
-                  access: 'unauthorized',
-                  auth: [],
-                  definitionId: 'provider-not-found',
-                },
-                data: {},
-              },
-            ),
-          );
-        } else {
-          dispatch(cardAction(ACTION_ERROR, { url: resourceUrl }, error));
-        }
+        handleResolvedLinkError(resourceUrl, error);
       } finally {
         isCompleted = true;
       }
     } else {
       dispatchAnalytics(resolvedEvent(definitionId));
       isCompleted = true;
+    }
+  }
+
+  function handleResolvedLinkResponse(resourceUrl: string, response: JsonLd) {
+    const nextDefinitionId = response && getDefinitionId(response);
+    const nextStatus = response ? getStatus(response) : 'fatal';
+
+    // Dispatch analytics.
+    if (nextStatus === 'resolved') {
+      dispatchAnalytics(resolvedEvent(nextDefinitionId));
+    } else {
+      dispatchAnalytics(unresolvedEvent(nextStatus, nextDefinitionId));
+      // If we require authorization & do not have an authFlow available,
+      // throw an error and render as a normal blue link.
+      if (nextStatus === 'unauthorized' && config.authFlow !== 'oauth2') {
+        handleResolvedLinkError(resourceUrl, ERROR_MESSAGE_OAUTH);
+        return;
+      }
+      if (nextStatus === 'fatal') {
+        handleResolvedLinkError(resourceUrl, ERROR_MESSAGE_FATAL);
+        return;
+      }
+    }
+
+    dispatch(cardAction(ACTION_RESOLVED, { url: resourceUrl }, response));
+  }
+
+  function handleResolvedLinkError(resourceUrl: string, error: any) {
+    const errorKind: string = error && error.kind;
+
+    // Handle FatalErrors (completely failed to resolve data).
+    if (errorKind === 'fatal') {
+      // If there's no previous data in the store for this URL, then bail
+      // out and let the editor handle fallbacks (returns to a blue link).
+      if (!details || status !== 'resolved') {
+        throw error;
+      }
+
+      // If we already have resolved data for this URL in the store, then
+      // simply fallback to the previous data.
+      if (status === 'resolved') {
+        dispatch(cardAction(ACTION_RESOLVED, { url: resourceUrl }, details));
+      }
+      // Handle AuthErrors (user did not have access to resource).
+    } else if (errorKind === 'auth') {
+      dispatch(
+        cardAction(
+          ACTION_RESOLVED,
+          { url: resourceUrl },
+          {
+            meta: {
+              visibility: 'restricted',
+              access: 'unauthorized',
+              auth: [],
+              definitionId: 'provider-not-found',
+            },
+            data: {},
+          },
+        ),
+      );
+    } else {
+      dispatch(cardAction(ACTION_ERROR, { url: resourceUrl }, error));
     }
   }